requirements.txt
streamlit
boto3
snowflake-connector-python